requirements.txt
pydantic-ai
litellm
rich
feedparser
httpx
python-dotenv
pyyaml
pydantic
